Remove GridScrollMixin
