+++ packages/react-data-grid/src/Grid.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 const createReactClass = require('create-react-class');
 const Header               = require('./Header');
 const Viewport             = require('./Viewport');
-const GridScrollMixin      = require('./GridScrollMixin');
 const DOMMetrics           = require('./DOMMetrics');
 const cellMetaDataShape    = require('./PropTypeShapes/CellMetaDataShape');
 require('../../../themes/react-data-grid-core.css');
@@ -65,7 +64,6 @@ const Grid = createReactClass({
   },
 
   mixins: [
-    GridScrollMixin,
     DOMMetrics.MetricsComputatorMixin
   ],
 
@@ -84,6 +82,50 @@ const Grid = createReactClass({
       position: 'relative',
       minHeight: this.props.minHeight
     };
+  },
+
+  _onScroll() {
+    if (this._scrollLeft !== undefined) {
+      this.header.setScrollLeft(this._scrollLeft);
+      if (this.viewport) {
+        this.viewport.setScrollLeft(this._scrollLeft);
+      }
+    }
+  },
+
+  onScroll(props) {
+    if (this._scrollLeft !== props.scrollLeft) {
+      this._scrollLeft = props.scrollLeft;
+      this._onScroll();
+    }
+  },
+
+  onHeaderScroll(e) {
+    let scrollLeft = e.target.scrollLeft;
+    if (this._scrollLeft !== scrollLeft) {
+      this._scrollLeft = scrollLeft;
+      this.header.setScrollLeft(scrollLeft);
+      let canvas = ReactDOM.findDOMNode(this.viewport.canvas);
+      canvas.scrollLeft = scrollLeft;
+      this.viewport.canvas.setScrollLeft(scrollLeft);
+    }
+  },
+
+  componentDidMount() {
+    this._scrollLeft = this.viewport ? this.viewport.getScroll().scrollLeft : 0;
+    this._onScroll();
+  },
+
+  componentDidUpdate() {
+    this._onScroll();
+  },
+
+  componentWillMount() {
+    this._scrollLeft = undefined;
+  },
+
+  componentWillUnmount() {
+    this._scrollLeft = undefined;
   },
 
   render(): ?ReactElement {
